requirements.txt
# 法印对照 · 依赖清单
# 安装: pip install -r requirements.txt

# ─── Web 框架 ────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0

# ─── XML 解析 ────────────────────────────────────────────────
lxml>=4.9.0

# ─── 简繁转换（搜索功能） ────────────────────────────────────
opencc-python-reimplemented>=0.1.7

# ─── HTTP 客户端（AI 释义 API 调用） ─────────────────────────
httpx>=0.25.0

# ─── AI 释义（SSE 流式调用，可选） ───────────────────────────
openai>=1.0.0

# ─── 用户词典 MDX 格式支持（可选） ───────────────────────────
mdict-utils>=1.3.0

# ─── 环境变量管理（可选但推荐） ──────────────────────────────
python-dotenv>=1.0.0
